Fixed tag logic for reversed tags

+++ src/main/scala/de/thatsich/minecraft/intellie/applied/aerodynamics/proxy/module/disassembler/DisassemblerItem.scala
@@ -65,8 +65,8 @@ extends BaseItem(new DisassemblerID, modid, log)
 		{
 			val energyUsage = this.getCurrentEnergyUsage(is).toInt
 			val multiplier = this.getCurrentChargeMultiplier(is).toInt
-			val speed = this.getCurrentMiningSpeed(is).toInt
-			val level = this.getCurrentMiningLevel(is).toInt
+			val speed = this.getCurrentMiningSpeed(is)
+			val level = this.getCurrentMiningLevel(is)
 			val damage = this.getCurrentDamageVsEntities(is).toInt
 
 			list.add(s"Stored Energy: $currentPower/$maxPower AE - $percent%")
